improved edit page, added and finished product details page, started login recovery page
basically touched up the editing stuff. it appears to be working and saving to the database. i'm unsure however, as to how i pass the information about the published variable as it is a date time class. i possibly need to change it to something thats just date without the time.

the product details page is basically finished. displays everything you need. however i still need to figure out how to display the name of the subgenre, since the way it's structured is quite complicated its hard to just call the name of the subgenre.

the login recovery page got it's page set up in the controller and a form added to take in an email but nothing else at this point until the login page is ready to go