login recovery almost works
turns out i forgot a few steps. started implementing them but when doing the controller action "LoginRecoveryReset" it will work the first pass, where it has no password. the model carries all the info over. so it asks the user for a new password. but when you hit submit for that reset, it is meant to run that action again but have a password this time, but the problem is that it does not keep the model information on the second passthrough.